LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/security.pattern.canvas.strokeStyle.sub-expected.txt

CONSOLE MESSAGE: line 32: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
security.pattern.canvas.strokeStyle.sub
Setting strokeStyle to a pattern of an unclean canvas makes the canvas origin-unclean
Actual output:

PASS Setting strokeStyle to a pattern of an unclean canvas makes the canvas origin-unclean 
